correção no endpoint de finalizar agenda

+++ APIDevSteam/Controllers/ItemCarrinhosController.cs
@@ -109,6 +109,21 @@ namespace APIDevSteam.Controllers
                 return NotFound();
             }
 
+            //Verificar se o carrinho existe
+            var carrinho = await _context.Carrinhos.FindAsync
+                (itemCarrinho.CarrinhoId);
+            if (carrinho == null)
+            {
+                return NotFound("Carrinho não encontrado");
+
+            }
+
+            //Remove o valor total do carrinho 
+            carrinho.ValorTotal -= itemCarrinho.ValorTotal;
+
+            //Verificar se o valor total do carrinho é menor que zero 
+
+
             _context.ItensCarrinhos.Remove(itemCarrinho);
             await _context.SaveChangesAsync();
